feat: cc-0082 logout function
