--- src/components/Login/api.ts
@@ -1,57 +1,0 @@
-import { environment } from "@/constants/environment";
-import axios from "axios";
-import useTokenService from "../../hooks/useTokenService";
-
-const { getLocalRefreshToken, getLocalAccessToken, updateLocalAccessToken } = useTokenService();
-
-const instance = axios.create({
-  baseURL: environment.apiBaseUrl,
-  headers: {
-    "Content-Type": "application/json",
-  },
-});
-
-instance.interceptors.request.use(
-  (config) => {
-    const token = getLocalAccessToken();
-    if (token) {
-      config.headers["x-access-token"] = token;
-    }
-    return config;
-  },
-  (error) => {
-    return Promise.reject(error);
-  }
-);
-
-instance.interceptors.response.use(
-  (res) => {
-    return res;
-  },
-  async (err) => {
-    const originalConfig = err.config;
-
-    if (originalConfig.url !== "/admin/login" && err.response) {
-      if (err.response.status === 401 && !originalConfig._retry) {
-        originalConfig._retry = true;
-
-        try {
-          const response = await instance.post("/admin/refresh", {
-            refreshToken: getLocalRefreshToken(),
-          });
-
-          const { accessToken } = response.data;
-          updateLocalAccessToken(accessToken);
-
-          return instance(originalConfig);
-        } catch (_error) {
-          return Promise.reject(_error);
-        }
-      }
-    }
-
-    return Promise.reject(err);
-  }
-);
-
-export default instance;
